requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
redis==5.0.1

# Message Queue
nats-py==2.6.0
aioredis==2.0.1

# Cryptography
cryptography==41.0.7
pycryptodome==3.19.0
pqcrypto==0.1.3
pycapnp==1.3.0

# Networking
scapy==2.5.0
requests==2.31.0
aiohttp==3.9.1
websockets==12.0

# OSINT
tweepy==4.14.0
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.16.0

# ML/AI
torch==2.1.1
transformers==4.35.2
numpy==1.24.3
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic-settings==2.1.0
rich==13.7.0
typer==0.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0

# Optional: Metasploit integration
pymetasploit3==1.1.0

# Optional: Hardware
pyserial==3.5
pyusb==1.2.1
